requirements.txt
pandas
sqlalchemy
snowflake-sqlalchemy
python-dotenv
dbt-core
dbt-snowflake